Fix QIcuTimeZonePrivate::hasDaylightTime() to catch historical DST
QTimeZone::hasDaylightTime() is documented to return true if a zone
has ever dabbled in DST, but the ICU implementation was only detecting
whether the zone currently engages in it. When the transitions API is
available (currently for version == 50; will soon change to >= 50),
use it to find out whether DST was ever in effect.

Task-number: QTBUG-99747
Change-Id: I4576de2e074dc57b54450b1e40a23358ccf0ef49
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit d9602efe8873083c7eb5171dc26b2ecdfa071841)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>